Solution for cf-1490D added
cf-1490D:
  - file: "non_overlapping_subproblem/divide_and_conquer/codeforces/
           F1490D/Solution.cpp"
    - another solution added
    - reason: different category

The reason to add this solution is to provide example for
non overlapping subproblems. Some problems does not fit into
the category of binary search or segment tree data structure.
They are just simple `decrease and conquer` or `divide and
conquer`. So, for them we need to create some category. That's
it.